fixed some bugs

+++ src/products/products.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Param,
@@ -7,7 +8,10 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { ProductsService } from './products.service.js';
-import { CreateProductDto } from './dto/create_product.dto.js';
+import {
+  CreateProductDto,
+  CreateSizeProductDto,
+} from './dto/create_product.dto.js';
 import { UpdateProductDto } from './dto/update_product.dto.js';
 import { Delete, Get, Patch, Post } from '@nestjs/common';
 import { AuthGuard } from '../auth/guards/auth.guard.js';
@@ -17,7 +21,7 @@ import { Roles } from '../auth/decorators/roles.decorator.js';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { ValidateImagePipe } from '../common/pipes.js';
 import { AwsService } from '../shared/aws.services.js';
-import { regexFile } from '../common/regex.js';
+import { allowedMimeTypes } from '../common/regex.js';
 
 @Controller('products')
 export class ProductsController {
@@ -33,7 +37,7 @@ export class ProductsController {
     FileInterceptor('image', {
       limits: { fileSize: 5 * 1024 * 1024 },
       fileFilter(req, file, callback) {
-        if (!file.mimetype.match(regexFile))
+        if (!allowedMimeTypes.includes(file.mimetype))
           return callback(new Error('File is invalid'), false);
         else callback(null, true);
       },
@@ -44,7 +48,16 @@ export class ProductsController {
     @UploadedFile(new ValidateImagePipe()) file: Express.Multer.File,
   ) {
     const imageUrl = await this.awsService.uploadFile(file);
-    return this.productsService.create({ ...dto, images: [imageUrl] });
+    if (typeof dto.sizes === 'string') {
+      const parsed = JSON.parse(dto.sizes) as CreateSizeProductDto[];
+
+      if (!Array.isArray(parsed)) {
+        throw new BadRequestException('sizes must be array');
+      }
+
+      dto.sizes = parsed;
+    }
+    return this.productsService.create(dto, imageUrl);
   }
 
   @Get()
@@ -57,9 +70,9 @@ export class ProductsController {
     return this.productsService.findOne(+id);
   }
 
-  @Get('all/:categoryName')
-  findProductsByCategory(@Param('categoryName') categoryName: string) {
-    return this.productsService.findByCategory(categoryName);
+  @Get('all/:categoryId')
+  findProductsByCategory(@Param('categoryId') categoryId: string) {
+    return this.productsService.findByCategory(+categoryId);
   }
 
   @UseGuards(AuthGuard, RolesGuard, CsrfGuard)
@@ -69,7 +82,7 @@ export class ProductsController {
     FileInterceptor('image', {
       limits: { fileSize: 5 * 1024 * 1024 },
       fileFilter(req, file, callback) {
-        if (!file.mimetype.match(regexFile))
+        if (!allowedMimeTypes.includes(file.mimetype))
           return callback(new Error('File is invalid'), false);
         else callback(null, true);
       },
@@ -84,6 +97,16 @@ export class ProductsController {
     if (!product) throw new Error('Product not found');
     const images = [...product.images];
 
+    if (typeof dto.sizes === 'string') {
+      const parsed = JSON.parse(dto.sizes) as CreateSizeProductDto[];
+
+      if (!Array.isArray(parsed)) {
+        throw new BadRequestException('sizes must be array');
+      }
+
+      dto.sizes = parsed;
+    }
+
     if (file) {
       for (const img of images) {
         await this.awsService.deleteFile(img);
@@ -94,15 +117,18 @@ export class ProductsController {
       images.push(uploadedImage);
     }
 
-    return this.productsService.update(+id, {
-      ...dto,
-      images,
-    });
+    return this.productsService.update(+id, dto, images);
   }
+
   @UseGuards(AuthGuard, RolesGuard, CsrfGuard)
   @Roles('ADMIN', 'USER')
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const product = await this.productsService.findOne(+id);
+    if (!product) throw new Error('Product not found');
+    for (const img of product.images) {
+      await this.awsService.deleteFile(img);
+    }
     return this.productsService.remove(+id);
   }
 }
